feat: recoil 템플릿 코드 추가
