+++ .eslintrc.json
@@ -57,7 +57,6 @@
         "allow": ["private-constructors"]
       }
     ],
-    "@typescript-eslint/no-unused-vars": "warn",
     "react/jsx-no-useless-fragment": "off",
     "react/react-in-jsx-scope": "off",
     "react/jsx-filename-extension": [1, { "extensions": [".tsx", ".jsx"] }],
