+++ src/features/auth/auth.hook.ts
@@ -1,6 +1,7 @@
 import { useRecoilState, useRecoilValue } from 'recoil';
 
-import { authState, getIsLogin } from './auth.model';
+import { authState } from './auth.atom';
+import { getIsLogin } from './auth.selector';
 
 export const useAuthState = () => useRecoilState(authState);
 export const useAuthIsLogin = () => useRecoilValue(getIsLogin);
