+++ src/features/auth/auth.selector.ts
@@ -1,0 +1,11 @@
+import { selector } from 'recoil';
+
+import { authState } from './auth.atom';
+
+export const getIsLogin = selector<boolean>({
+  key: 'authState/isLogin',
+  get: ({ get }) => {
+    const auth = get(authState);
+    return auth.isLogin;
+  },
+});
